requirements.txt
torch>=2.1.0
transformers>=4.40.0
datasets>=2.19.0
peft>=0.10.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
gradio>=4.0.0
evaluate>=0.4.0
tqdm>=4.66.0
spacy>=3.7.0
huggingface_hub>=0.23.0
python-dotenv>=1.0.0
streamlit>=1.35.0